@@ -1,6 +1,6 @@
 {
   "name": "@amberfly/crankshaft",
-  "version": "0.1.8",
+  "version": "0.2.0",
   "description": "",
   "main": "./server/index.js",
   "type": "module",
@@ -17,14 +17,15 @@
   "dependencies": {
     "nes.css": "^2.3.0",
     "npm-run-all": "^4.1.5",
-    "react": "^17.0.2",
-    "react-dom": "^17.0.2",
+    "react": "^18.3.1",
+    "react-dom": "^18.3.1",
     "react-fetch-hook": "^1.8.5",
+    "react-tabs": "^6.1.0",
     "ts-loader": "^9.2.8"
   },
   "devDependencies": {
-    "@types/react": "^17.0.11",
-    "@types/react-dom": "^17.0.7",
+    "@types/react": "^18.3.1",
+    "@types/react-dom": "^18.3.1",
     "@typescript-eslint/eslint-plugin": "^4.27.0",
     "@typescript-eslint/parser": "^4.27.0",
     "copy-webpack-plugin": "^9.0.0",
