@@ -1,9 +1,10 @@
 import React from 'react';
 import { Server } from './Server';
-import '../styles.scss';
 import { Header } from './Header';
 import { BedrockVersion, ServerState } from '../../interfaces/types';
 import useFetch from 'react-fetch-hook';
+import { Tab, TabList, TabPanel, Tabs } from 'react-tabs';
+import '../styles.scss';
 
 export const App: React.FC = () => {
     const version = useFetch<BedrockVersion[]>('/bedrock/versions');
@@ -18,9 +19,12 @@ export const App: React.FC = () => {
                         <p>Loading...</p>
                     </div>
                 ) : (
-                    <>
-                    {servers.data?.map((s: ServerState, index: number) => <Server key={index} installers={version.data} index={index} /> )}
-                    </>
+                    <Tabs>
+                        <TabList>
+                            { servers.data?.map((s: ServerState, index: number) => <Tab  className="nes-container is-rounded is-dark" key={index}>{index}</Tab> ) }
+                        </TabList>
+                            { servers.data?.map((s: ServerState, index: number) => <TabPanel key={index}><Server key={index} installers={version.data} index={index} /></TabPanel> ) }
+                    </Tabs>
                 )}
             </div>
         </>
